Add test coverage for NetX::HTTPUnix.new
Puppet's HTTP client connection handling uses Net::HTTP.new as the
interface.  This is a problem because only the start class method has
test coverage.  This patch addresses the problem by adding test coverage
for the new class method to make sure the library is compatible with
Puppet.

This patch also updates the README