Fix incorrect HOST request header causing 400 errors
Without this patch nginx returns a 400 error response for requests that
otherwise are 200 OK over TCP connections.  The specific error message
is:

    2014/11/25 21:47:24 [info] 19036#0: *1 client sent invalid host
    header while reading client request headers, client: unix:, server:
    localhost, request: "POST /command-api HTTP/1.1", host:
    "/var/run/command-api.sock:"

This patch addresses the problem by storing a valid hostname of
'localhost' in the @address instance variable rather than the socket
path.  This change allows the HTTP client to construct a HOST request
header nginx will accept.  The unix domain socket specific connection
method uses the @socket_path instance variable instead of relying on the
behavior of conn_address() which in turn used @address.