(maint) improve test suite to work on modern rubies
Without this change, tests would fail with Errno::ECONNRESET or Errno::EPIPE as the server would close the connection without reading all of the client's messages. Trying to actually read the request lead to lock-ups. Since this doesn't actually have to read anything, `close_write` works well enough for this.